--- a/lib/core.js
+++ b/lib/core.js
@@ -35,12 +35,12 @@ function Promise(fn) {
       deferreds.push(deferred)
       return
     }
+    var cb = state ? deferred.onFulfilled : deferred.onRejected
+    if (cb === null) {
+      (state ? deferred.resolve : deferred.reject)(value)
+      return
+    }
     asap(function() {
-      var cb = state ? deferred.onFulfilled : deferred.onRejected
-      if (cb === null) {
-        (state ? deferred.resolve : deferred.reject)(value)
-        return
-      }
       var ret
       try {
         ret = cb(value)