I did a little benchmarking and I believe these improvements together result in somewhere between 2% and 5% speed up.  they don't have any other observable impact, so I think they're a net win (albeit a small one)

Benchmark on master:

```
bluebird x 84.93 ops/sec ±1.97% (81 runs sampled)
es6-promises x 54.57 ops/sec ±2.42% (68 runs sampled)
legendary x 53.10 ops/sec ±2.87% (66 runs sampled)
lie x 85.25 ops/sec ±2.28% (71 runs sampled)
native x 72.57 ops/sec ±2.17% (75 runs sampled)
native-promise-only x 72.93 ops/sec ±1.46% (71 runs sampled)
promise x 62.81 ops/sec ±3.93% (73 runs sampled)
q x 31.55 ops/sec ±2.87% (58 runs sampled)
vow x 81.39 ops/sec ±1.02% (78 runs sampled)
```
Benchmark using this branch:

```
bluebird x 83.81 ops/sec ±2.45% (81 runs sampled)
es6-promises x 53.03 ops/sec ±2.73% (66 runs sampled)
legendary x 53.76 ops/sec ±2.93% (67 runs sampled)
lie x 82.69 ops/sec ±2.10% (70 runs sampled)
native x 73.44 ops/sec ±1.25% (72 runs sampled)
native-promise-only x 71.81 ops/sec ±1.51% (70 runs sampled)
promise x 65.19 ops/sec ±2.90% (69 runs sampled)
q x 31.61 ops/sec ±2.63% (60 runs sampled)
vow x 79.01 ops/sec ±2.84% (76 runs sampled)
```

These figures are still sadly prety bad for us.